move annotation functions to new file
